feat: forgot collect flownode clusterinfo handler
