+++ src/meta-srv/src/metasrv/builder.rs
@@ -46,15 +46,16 @@ use crate::flow_meta_alloc::FlowPeerAllocator;
 use crate::greptimedb_telemetry::get_greptimedb_telemetry_task;
 use crate::handler::check_leader_handler::CheckLeaderHandler;
 use crate::handler::collect_cluster_info_handler::{
-    CollectDatanodeClusterInfoHandler, CollectFrontendClusterInfoHandler,
+    CollectDatanodeClusterInfoHandler, CollectFlownodeClusterInfoHandler,
+    CollectFrontendClusterInfoHandler,
 };
 use crate::handler::collect_stats_handler::CollectStatsHandler;
+use crate::handler::extract_stat_handler::ExtractStatHandler;
 use crate::handler::failure_handler::RegionFailureHandler;
 use crate::handler::filter_inactive_region_stats::FilterInactiveRegionStatsHandler;
 use crate::handler::keep_lease_handler::{DatanodeKeepLeaseHandler, FlownodeKeepLeaseHandler};
 use crate::handler::mailbox_handler::MailboxHandler;
 use crate::handler::on_leader_start_handler::OnLeaderStartHandler;
-use crate::handler::persist_stats_handler::PersistStatsHandler;
 use crate::handler::publish_heartbeat_handler::PublishHeartbeatHandler;
 use crate::handler::region_lease_handler::RegionLeaseHandler;
 use crate::handler::response_header_handler::ResponseHeaderHandler;
@@ -361,9 +362,10 @@ impl MetasrvBuilder {
                 group.add_handler(FlownodeKeepLeaseHandler).await;
                 group.add_handler(CheckLeaderHandler).await;
                 group.add_handler(OnLeaderStartHandler).await;
-                group.add_handler(CollectStatsHandler).await;
+                group.add_handler(ExtractStatHandler).await;
                 group.add_handler(CollectDatanodeClusterInfoHandler).await;
                 group.add_handler(CollectFrontendClusterInfoHandler).await;
+                group.add_handler(CollectFlownodeClusterInfoHandler).await;
                 group.add_handler(MailboxHandler).await;
                 group.add_handler(region_lease_handler).await;
                 group.add_handler(FilterInactiveRegionStatsHandler).await;
@@ -373,7 +375,7 @@ impl MetasrvBuilder {
                 if let Some(publish_heartbeat_handler) = publish_heartbeat_handler {
                     group.add_handler(publish_heartbeat_handler).await;
                 }
-                group.add_handler(PersistStatsHandler::default()).await;
+                group.add_handler(CollectStatsHandler::default()).await;
                 group
             }
         };
